Added Symphony pre-requisites
Added info on things you need to do in Symphony before you can use Terraform against it. These include creating a dedicated VPC-enabled project for use by Terraform/AWS API in general, associating a shared edge network with that project, and creating access and secret keys for that project. Also included in to on how to obtain an image's AWS ID from the Symphony GUI.